set up build scripts using webpack, actually add all needed dependencies so now it actually works
