+++ .eslintrc
@@ -1,0 +1,15 @@
+{
+  "extends": [ "eslint:recommended", "plugin:react/recommended" ],
+  "env": {
+    “node”: true,
+    “browser”: true,
+  },
+  "parserOptions": {
+      ecmaVersion: 6,
+      ecmaFeatures: {
+          jsx: true,
+      },
+      sourceType: "module"
+  },
+  "plugins": [ "react" ]
+}
