+++ .eslintrc.js
@@ -1,0 +1,3 @@
+module.exports = {
+	parser: "babel-eslint",
+};
