--- index.html
@@ -1,180 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <meta charset="UTF-8">
-  <title></title>
-</head>
-<body>
-<style type="text/css">
-        html, body {
-         
-        }
-
-        #root {
-            min-height: 100%; 
-        }
-    </style>
-    <script id="vertex_shader" type="x-shader/x-vertex">
-
-varying vec2 vUv;
-
-void main() {
-
-    vUv = uv;
-    
-    gl_Position = projectionMatrix * modelViewMatrix * vec4( position, 1.0 );
-    
-}
-
-</script>
-
-<script id="fragment_shader_0" type="x-shader/x-fragment">
-
-uniform vec3 color;
-uniform float time;
-varying vec2 vUv;
-
-void main() {
-
-    vec4 fragColor;
-
-      fragColor = vec4(color, 1.0);
-    
-    
-    gl_FragColor = fragColor;
-}
-
-</script>
-<script id="fragment_shader_1" type="x-shader/x-fragment">
-
-uniform vec3 color;
-uniform sampler2D texture1;
-uniform sampler2D imgtexture0;
-uniform sampler2D imgtexture1;
-uniform sampler2D imgtexture2;
-uniform sampler2D imgtexture3;
-uniform sampler2D imgtexture4;
-
-uniform float xMin1;
-uniform float xMax1;
-uniform float yMin1;
-uniform float yMax1;
-uniform float zMin1;
-uniform float zMax1;
-
-uniform float xMin2;
-uniform float xMax2;
-uniform float yMin2;
-uniform float yMax2;
-uniform float zMin2;
-uniform float zMax2;
-
-uniform float xMin3;
-uniform float xMax3;
-uniform float yMin3;
-uniform float yMax3;
-uniform float zMin3;
-uniform float zMax3;
- 
-uniform float xMin4;
-uniform float xMax4;
-uniform float yMin4;
-uniform float yMax4;
-uniform float zMin4;
-uniform float zMax4;
-
-
-
-uniform float time;
-varying vec2 vUv;
-
-void main() {
-
-
-    vec4 tColor;
-    vec4 tColor2;
-    vec4 fragColor;
-
-      vec4 col = vec4(color, 1.0);
-      tColor = texture2D( texture1, vUv )*col;
-      
-      //pixel is black/
-      /*
-      if(abs(tColor.x-tColor.y)<.01&&abs(tColor.y-tColor.z)<.01&&tColor.x<.1){
-        fragColor = texture2D( imgtexture0, vUv);
-      }
-      //white
-      
-      else if(abs(tColor.x-tColor.y)<.01&&abs(tColor.y-tColor.z)<.01&&tColor.x>=.1){
-        fragColor = texture2D( imgtexture0, vUv);
-      }
-     
-      //red
-      
-      else if(abs(tColor.x-tColor.y)>.9&&tColor.y<.1&&tColor.z<.1){
-        fragColor = texture2D( imgtexture1, vUv);
-      }
-       
-    //blue
-      else if(tColor.z>.9&&tColor.x<.1&&tColor.y<.1){
-        fragColor = texture2D( imgtexture1, vUv);
-      }
-      //green
-      else{
-        fragColor = texture2D(imgtexture1, vUv);
-      }
-    */
-    if(tColor.x<.5&&tColor.y<.5&&tColor.z<.5){
-      fragColor = texture2D(imgtexture0, vUv);
-    }
-    else if(tColor.x<.5&&tColor.y<.5){
-      fragColor = texture2D(imgtexture1, vUv);
-    }
-    else if(tColor.x>.5&&tColor.z<0.5){
-      fragColor=texture2D(imgtexture2, vUv);
-    }
-    else if(tColor.z>0.5&&tColor.y>0.5){
-      fragColor=texture2D(imgtexture3, vUv);
-    }
-    else{
-      fragColor=texture2D(imgtexture4, vUv);
-    }
-   
-    gl_FragColor = fragColor;
-}
-
-</script>
-<script id="fragment_shader_2" type="x-shader/x-fragment">
-
-uniform vec3 color;
-uniform sampler2D texture1;
-uniform sampler2D texture2;
-uniform float time;
-varying vec2 vUv;
-
-void main() {
-
-    vec4 tColor;
-    vec4 tColor2;
-    vec4 fragColor;
-
-   
-  
-      tColor = texture2D( texture1, vUv );
-      tColor2 = texture2D(texture2, vUv);
-      fragColor = vec4(tColor.x*tColor2.x*color.x, tColor.y*tColor2.y*color.y, tColor.z*tColor2.z*color.z, 1.0);
-    gl_FragColor = fragColor;
-}
-
-</script>
-
-<input type="checkbox" id="record"/>record
- <div id="root"></div>
-
- <div id="three-box"></div>
-
-<script>window.$ = window.jQuery = require('jquery');</script>
-<script src="app/app.jsx"></script>
-<script src="react-three/build/react-three.js"></script>
-</body>
-</html>
