+++ package.json
@@ -11,24 +11,36 @@
     ]
   },
   "scripts": {
-    "start": "electron es6-init.js"
+    "start": "electron .",
+    "build": "webpack --mode development"
   },
   "dependencies": {
     "app-root-path": "^2.0.1",
     "babel-preset-es2015": "^6.18.0",
-    "babel-preset-react": "^6.16.0",
     "electron-compile": "^4.0.2",
     "file-saver": "^1.3.3",
     "fs": "0.0.1-security",
     "fs-web": "^1.0.1",
     "jquery": "^3.1.1",
+    "path": "^0.12.7",
     "react": "^15.4.1",
     "react-color": "^2.9.0",
     "react-dom": "^15.4.1",
     "react-three": "^0.9.5",
-    "three": "^0.74.0"
+    "three": "^0.83.0"
   },
   "devDependencies": {
-    "electron-compilers": "^4.0.5"
+    "babel-core": "^6.26.3",
+    "babel-eslint": "^10.0.3",
+    "babel-loader": "^7.1.5",
+    "babel-preset-env": "^1.7.0",
+    "babel-preset-react": "^6.24.1",
+    "electron-compilers": "^4.0.5",
+    "eslint": "^6.6.0",
+    "html-webpack-plugin": "^3.2.0",
+    "webpack": "^4.41.2",
+    "webpack-cli": "^3.3.10",
+    "webpack-dev-server": "^3.9.0",
+    "webpack-shell-plugin": "^0.5.0"
   }
 }
