+++ src/fileProcess.jsx
@@ -12,7 +12,7 @@ export default class FileProcess extends React.Component{
 
 	drop(e) {
 		e.preventDefault();
-		var media = [];
+		const media = [];
     	for (let f of e.dataTransfer.files) {
       		media.push(f.path);
     	}
@@ -22,13 +22,13 @@ export default class FileProcess extends React.Component{
 	render() {
 		return(
 			<div id="holder"
-			onDragOver = {(e)=>this.dragOver(e)}
-			onDragEnter = {(e)=>this.dragOver(e)}
-			onDragLeave = {(e)=>this.dragLeave(e)}
-			onDragEnd = {(e)=>this.dragLeave(e)}
-			onDrop = {(e)=>this.drop(e)}
+				onDragOver = { (e)=>this.dragOver(e )}
+				onDragEnter = { (e)=>this.dragOver(e) }
+				onDragLeave = { (e)=>this.dragLeave(e) }
+				onDragEnd = { (e)=>this.dragLeave(e) }
+				onDrop = { (e)=>this.drop(e) }
 			>	
-				<div id="box__input" style={this.props.style.display}>
+				<div id="box__input" style={ this.props.style.display } >
 					<strong>Choose a file </strong>
 					 or drag it here
 				</div>
